requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.25.1

# Linting
black==23.1.0
isort==5.12.0
autoflake==2.0.1